+++ test/config.js
@@ -1,8 +1,0 @@
-module.exports = {
-    file: 'js',
-    name: 'name',
-    db: {
-        host: 'url',
-        user: 'User'
-    }
-}