+++ CHANGELOG.md
@@ -4,6 +4,25 @@
 All notable changes to this project will be documented in this file.
 <!--- END HEADER -->
 
+## [0.9.2](https://gitlab-p2.autobid.de/t.budzynski/node-config/compare/0.9.1...0.9.2) (2022-02-01)
+### Bug Fixes
+
+* Array config changed to object value ([cc0bc9](https://gitlab-p2.autobid.de/t.budzynski/node-config/commit/cc0bc9d7ee325cdc604ec8e3443ec7162260bc6e))
+
+### Tests
+
+* New test units ([f29d7d](https://gitlab-p2.autobid.de/t.budzynski/node-config/commit/f29d7de06a47cf3d61f58577d367a691e44e53fd))
+
+### Documentation
+
+
+##### Readme
+
+* Fix ([a59a64](https://gitlab-p2.autobid.de/t.budzynski/node-config/commit/a59a64a8e06def434f51bd4d4b4a8cdc279f8d13))
+
+
+---
+
 ## [0.9.1](https://gitlab-p2.autobid.de/t.budzynski/node-config/compare/0.9.0...0.9.1) (2021-11-29)
 ### Bug Fixes